payment controller and service made

+++ src/main/java/p/jesse/accountor/entities/Expense.java
@@ -3,6 +3,7 @@ package p.jesse.accountor.entities;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -11,7 +12,7 @@ public class Expense extends Payment {
     @ManyToOne
     private User user;
 
-    public Expense(double amount, String description, LocalDate createdAt, boolean isContinuous, Category category, User user) {
+    public Expense(BigDecimal amount, String description, LocalDate createdAt, boolean isContinuous, Category category, User user) {
         super(amount, description, createdAt, isContinuous, category);
         this.user = user;
     }
