+++ src/main/java/p/jesse/accountor/entities/Income.java
@@ -3,6 +3,7 @@ package p.jesse.accountor.entities;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -13,7 +14,7 @@ public class Income extends Payment {
     @ManyToOne
     private User user;
 
-    public Income(double amount, String description, LocalDate createdAt, boolean isContinuous, Category category, String source, User user) {
+    public Income(BigDecimal amount, String description, LocalDate createdAt, boolean isContinuous, Category category, String source, User user) {
         super(amount, description, createdAt, isContinuous, category);
         this.source = source;
         this.user = user;
